+++ django_cas_ng/__init__.py
@@ -7,6 +7,7 @@ __all__ = []
 _DEFAULTS = {
     'CAS_ADMIN_PREFIX': None,
     'CAS_EXTRA_LOGIN_PARAMS': None,
+    'CAS_RENEW': False,
     'CAS_IGNORE_REFERER': False,
     'CAS_LOGOUT_COMPLETELY': True,
     'CAS_REDIRECT_URL': '/',