+++ docs/changelog.rst
@@ -6,6 +6,11 @@ Listed are the high-level, notable changes for each django-cas-ng release.
 Backwards incompatible changes or other upgrade issues are also described
 here. For additional detail, read the complete `commit history`_.
 
+**django-cas-ng 3.2.0** ``[2014-xx-xx]``
+
+* Add CAS_RENEW setting to enforce CAS renew feature. Default is False.
+
+
 **django-cas-ng 3.1.0** ``[2014-05-25]``
    * Support Django 1.5 custom user model.